refactor: list rendering, simplify draw function
Removed the toggling of the hidden class on the ul, since the same
effect can be achieved by just clearing the list and redrawing it when
needed. By also removing the special case of drawing one item, all of
the list rendering code becomes simpler and can be used for all of
the event handlers.